requirements.txt
0x-contract-addresses==3.0.0
0x-contract-artifacts==3.0.0
0x-json-schemas==2.1.0
0x-order-utils==4.0.1
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==2.6.0
bleach==5.0.1
bs4==0.0.1
cbor2==5.4.3
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.0
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
cytoolz==0.12.0
dacite==1.6.0
daphne==3.0.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1
entrypoints==0.4
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
executing==0.10.0
fastjsonschema==2.16.1
frozenlist==1.3.1
h11==0.13.0
hexbytes==0.2.3
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipfshttpclient==0.8.0a2
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
lru-dict==1.1.8
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
msgpack==1.0.4
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
nbclient==0.6.6
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
netaddr==0.8.0
notebook==6.4.12
outcome==1.2.0
packaging==21.3
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.3
pickleshare==0.7.5
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pyee==9.0.4
Pygments==2.13.0
pymerkle==2.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
pywin32==304
pywinpty==2.0.7
pyzmq==23.2.1
qtconsole==5.3.1
QtPy==2.2.0
requests==2.28.1
rlp==2.0.1
selenium==4.4.3
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.4.0
stringcase==1.2.0
terminado==0.15.0
thirdweb-contract-wrappers==2.0.4
thirdweb-eth-account==0.6.6
thirdweb-sdk==2.0.18
tinycss2==1.1.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.11
varint==1.0.2
wcwidth==0.2.5
web3==5.27.0
webencodings==0.5.1
websockets==9.1
widgetsnbextension==4.0.2
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.1
zope.interface==5.4.0
